Separate milmove into files by sub command
